fixed a colors glitch in find box, when interacting with console box (PopupOpening not called which switching panes)
dirty bit is now cleared if you undo all the way to the beginning

fixed issue where pastes of large block comments might not
be colored correctly until you scroll up enough for the start
of the comment to become visible.

there was an annoying syntax-coloring bug where numbers could
temporarily lose their color when inserted next to a brace; finally
got around to tracking it down, so it is gone now, yay!

Drag n' Drop into the window works now

the BFilePanel was previously not dismissed after an "Open", only after
a "Save", which meant it stuck around until the next time the "Open" was
done, at which point it would be dismissed then. "ccache" for some reason liked
to occasionally crash this hidden file panel, taking Sisong down with it. Unsure why the BFilePanel is crashing, but it's delayed dismissal is fixed, so there should be no problems building Sisong from within Sisong.

re-organization of "Misc.cpp"; moving some of this functionality
back into the common Misc.cpp, where it was originally
(some of SS's Misc.cpp was copied out of a common Misc.cpp shared
with my other projects, but the common Misc previously wouldn't
compile on Haiku).



git-svn-id: svn://svn.code.sf.net/p/five75/code/sisong@6 0a535a1b-ecd5-406b-889c-d3337838b04b